+++ sources/scripts/client/gui/aslainMenu/__init__.py
@@ -6,7 +6,7 @@ __credits__ = ['Renat Iliev (izeberg) - original author',
 	'Andrii Andrushchyshyn (poliroid) - original contributor',
 	'Paul Ekshmidt (Kurzdor) - original maintainer']
 __license__ = 'CC BY-NC-SA 4.0'
-__version__ = '1.4.1'
+__version__ = '1.5.0'
 __maintainer__ = 'Aslain'
 __doc__ = 'https://github.com/Aslain/modssettingsapi'
 
@@ -125,6 +125,5 @@ class _ModsSettingsApi(IModsSettingsApi):
 		:return: bool
 		"""
 		return self.__instance.checkKeyset(keyset)
-
 
 g_modsSettingsApi = ModsSettingsApi()